requirements.txt
crewai[tools]==0.28.8
exa-py==1.15.6
langchain-groq==0.1.3
langchain==0.1.20
python-dotenv==1.0.0
langsmith==0.1.147
IPython==8.16.1
rich==13.9.4
tqdm==4.67.1
requests==2.32.5
numpy==1.26.4
pandas==2.1.1